+++ src/main/java/com/htmake/htbot/discord/commands/dungeon/data/FieldDungeonStatus.java
@@ -3,19 +3,20 @@ package com.htmake.htbot.discord.commands.dungeon.data;
 import lombok.*;
 
 import java.util.List;
+import java.util.Map;
 
 @Getter
 @Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class DungeonStatus {
+public class FieldDungeonStatus {
 
     private String name;
 
     private int stage;
 
-    private List<DungeonMonster> dungeonMonsterList;
+    private Map<Integer, DungeonMonster> monsterByStage;
 
     private DungeonPlayer dungeonPlayer;
